+++ src/app/Library/Fields/RadiosField.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace LaraWhale\Cms\Library\Fields;
+
+class RadiosField extends MultiCheckableField
+{
+    //
+}
