+++ tests/Support/Entries/all_types.php
@@ -20,6 +20,17 @@ return [
             ],
         ],
         [
+            'key' => 'radios',
+            'type' => 'radios',
+            'config' => [
+                'list' => [
+                    'item 1',
+                    'item 2',
+                    'item 3',
+                ],
+            ],
+        ],
+        [
             'key' => 'route',
             'type' => config('cms.fields.route_field_type'),
         ],
